Simplify implementation
In the presentation of the paper, nodes with two parents are tracked
with a second parent. The reason being that the children of a node are
not list contained in the node, but a pointer to the head of the
list (each node points to its next sibling). In a node w with two parents
u and v, both u and v may point (indirectly) to w, but then the next
sibling of w seem to belong both to the children of u and those of
v (which is not always possible). Having a pointer as the second parent
disambiguate that, as w is _the last children_ of his second parent.

But since we use actual list in this implementation (at least for now),
we only need a simple reference count.