Replace polymorphic compare by Ord.compare
The comparison function provided in the functor argument Ord was actually not used in the code, which was instead using the polymorphic comparison.

It might be useful to include some automated tests which use a non-standard ordering: for instance, integers with the reverse ordering to make a max heap. This would easily watch this bug.